feat: Export stub driver
