+++ src/qibolab_qblox/__init__.py
@@ -1,1 +1,7 @@
+from . import _core
+from ._core import *
+
+__all__ = []
+__all__ += _core.__all__
+
 __version__ = "0.0.3"
